requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.0
language-tool-python>=2.7.1
textstat>=0.7.3
pandas>=2.0.0
gradio>=4.0.0
tqdm>=4.66.0
